extract out //c10/core:alloc_cpu
Differential Revision: [D33329688](https://our.internmc.facebook.com/intern/diff/D33329688/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D33329688/)!

[ghstack-poisoned]
