+++ c10/BUILD.bazel
@@ -78,6 +78,7 @@ cc_library(
         ":headers",
         "@fmt",
         "//c10/core:ScalarType",
+        "//c10/core:alloc_cpu",
         "//c10/core:base",
         "//c10/util:TypeCast",
         "//c10/util:base",
