+++ c10/core/CPUAllocator.cpp
@@ -1,105 +1,17 @@
 #include <c10/core/CPUAllocator.h>
 #include <c10/core/DeviceType.h>
 #include <c10/core/alignment.h>
+#include <c10/core/impl/alloc_cpu.h>
 #include <c10/mobile/CPUCachingAllocator.h>
 #include <c10/mobile/CPUProfilingAllocator.h>
-#include <c10/util/irange.h>
 
-// TODO: rename flags to C10
+// TODO: rename flag to C10
 C10_DEFINE_bool(
     caffe2_report_cpu_memory_usage,
     false,
     "If set, print out detailed memory usage");
 
-C10_DEFINE_bool(
-    caffe2_cpu_allocator_do_zero_fill,
-    false,
-    "If set, do memory zerofilling when allocating on CPU");
-
-C10_DEFINE_bool(
-    caffe2_cpu_allocator_do_junk_fill,
-    false,
-    "If set, fill memory with deterministic junk when allocating on CPU");
-
 namespace c10 {
-
-void memset_junk(void* data, size_t num) {
-  // This garbage pattern is NaN when interpreted as floating point values,
-  // or as very large integer values.
-  static constexpr int32_t kJunkPattern = 0x7fedbeef;
-  static constexpr int64_t kJunkPattern64 =
-      static_cast<int64_t>(kJunkPattern) << 32 | kJunkPattern;
-  int32_t int64_count = num / sizeof(kJunkPattern64);
-  int32_t remaining_bytes = num % sizeof(kJunkPattern64);
-  int64_t* data_i64 = reinterpret_cast<int64_t*>(data);
-  for (const auto i : c10::irange(int64_count)) {
-    data_i64[i] = kJunkPattern64;
-  }
-  if (remaining_bytes > 0) {
-    memcpy(data_i64 + int64_count, &kJunkPattern64, remaining_bytes);
-  }
-}
-
-void* alloc_cpu(size_t nbytes) {
-  if (nbytes == 0) {
-    return nullptr;
-  }
-  // We might have clowny upstream code that tries to alloc a negative number
-  // of bytes. Let's catch it early.
-  CAFFE_ENFORCE(
-      ((ptrdiff_t)nbytes) >= 0,
-      "alloc_cpu() seems to have been called with negative number: ",
-      nbytes);
-
-  // NOLINTNEXTLINE(cppcoreguidelines-init-variables)
-  void* data;
-#ifdef __ANDROID__
-  data = memalign(gAlignment, nbytes);
-#elif defined(_MSC_VER)
-  data = _aligned_malloc(nbytes, gAlignment);
-#else
-  int err = posix_memalign(&data, gAlignment, nbytes);
-  if (err != 0) {
-    CAFFE_THROW(
-        "DefaultCPUAllocator: can't allocate memory: you tried to allocate ",
-        nbytes,
-        " bytes. Error code ",
-        err,
-        " (",
-        strerror(err),
-        ")");
-  }
-#endif
-
-  CAFFE_ENFORCE(
-      data,
-      "DefaultCPUAllocator: not enough memory: you tried to allocate ",
-      nbytes,
-      " bytes.");
-
-  // move data to a thread's NUMA node
-  NUMAMove(data, nbytes, GetCurrentNUMANode());
-  CHECK(
-      !FLAGS_caffe2_cpu_allocator_do_zero_fill ||
-      !FLAGS_caffe2_cpu_allocator_do_junk_fill)
-      << "Cannot request both zero-fill and junk-fill at the same time";
-  if (FLAGS_caffe2_cpu_allocator_do_zero_fill) {
-    memset(data, 0, nbytes);
-  } else if (FLAGS_caffe2_cpu_allocator_do_junk_fill) {
-    memset_junk(data, nbytes);
-  }
-
-  return data;
-}
-
-void free_cpu(void* data) {
-#ifdef _MSC_VER
-  _aligned_free(data);
-#else
-  // NOLINTNEXTLINE(cppcoreguidelines-no-malloc)
-  free(data);
-#endif
-}
 
 struct C10_API DefaultCPUAllocator final : at::Allocator {
   DefaultCPUAllocator() = default;
