+++ c10/core/CPUAllocator.h
@@ -5,12 +5,9 @@
 
 #include <c10/core/Allocator.h>
 #include <c10/util/Logging.h>
-#include <c10/util/numa.h>
 
 // TODO: rename to c10
 C10_DECLARE_bool(caffe2_report_cpu_memory_usage);
-C10_DECLARE_bool(caffe2_cpu_allocator_do_zero_fill);
-C10_DECLARE_bool(caffe2_cpu_allocator_do_junk_fill);
 
 namespace c10 {
 
@@ -18,14 +15,6 @@ using MemoryDeleter = void (*)(void*);
 
 // A helper function that is basically doing nothing.
 C10_API void NoDelete(void*);
-
-// Fill the data memory region of num bytes with a particular garbage pattern.
-// The garbage value is chosen to be NaN if interpreted as floating point value,
-// or a very large integer.
-C10_API void memset_junk(void* data, size_t num);
-
-C10_API void* alloc_cpu(size_t nbytes);
-C10_API void free_cpu(void* data);
 
 // A simple struct that is used to report C10's memory allocation and
 // deallocation status to the profiler
